fix(desktop): address review feedback on PR #4675
Functional fixes:
- ProjectLocationSection: the "Change location" pencil button used to
  open the local native folder picker even when the project was set
  up on a remote host, which would have produced an invalid
  relocation path. Now: on local targets it still opens the native
  picker; on remote targets it opens the RemotePathPicker against
  the remote host's filesystem. The relocate flow (conflict
  precheck + relocate confirm dialog) is unchanged.
- V2ScriptsEditor/ScriptField: add `pb-8 pr-20` to the textarea so
  the absolute-positioned Import button at the bottom-right can no
  longer overlap typed content.
- ClickablePath: the truncate span needs `min-w-0` when its parent
  is a flex container, otherwise the ellipsis never kicks in.

Structural fixes (one-component-per-file rule from AGENTS.md):
- Extract `SettingsRow` from V2ProjectSettings.tsx into its own
  SettingsRow/ folder with an index re-export. V2ProjectSettings.tsx
  imports it.
- Extract `ScriptField` from V2ScriptsEditor.tsx into a co-located
  components/ScriptField/ folder. V2ScriptsEditor.tsx imports it.
  (Co-located under V2ScriptsEditor since it's only used by that
  editor.)
