+++ apps/desktop/src/renderer/routes/_authenticated/settings/v2-project/$projectId/components/V2ProjectSettings/components/SettingsRow/index.ts
@@ -1,0 +1,1 @@
+export { SettingsRow } from "./SettingsRow";
