+++ apps/desktop/src/renderer/routes/_authenticated/settings/v2-project/$projectId/components/V2ProjectSettings/components/V2ScriptsEditor/components/ScriptField/index.ts
@@ -1,0 +1,1 @@
+export { ScriptField } from "./ScriptField";
